refactor: move type to different file
