+++ proxmox/data_qemu_agent.go
@@ -1,0 +1,12 @@
+package proxmox
+
+import (
+	"net"
+)
+
+type AgentNetworkInterface struct {
+	MACAddress  string
+	IPAddresses []net.IP
+	Name        string
+	Statistics  map[string]int64
+}
